# Add HashMap.insertWithKey.