--- a/Data/HashMap/Lazy.hs
+++ b/Data/HashMap/Lazy.hs
@@ -42,6 +42,7 @@ module Data.HashMap.Lazy
     , (!)
     , insert
     , insertWith
+    , insertWithKey
     , delete
     , adjust
     , update