This function is present in the container's Map interface, [Map.insertWithKey](https://hackage.haskell.org/package/containers-0.5.10.2/docs/Data-Map-Strict.html#v:insertWithKey), this brings more consistency to the `HashMap` interface so that it can more easily be used as a drop in replacement for `Data.Map` if desired.

This partially addresses https://github.com/tibbe/unordered-containers/issues/172.