@m-renaud 2018-01-23
Currently working on benchmarks, there aren't any existing benchmarks for `insertWith` so I'll be adding those as well.