@m-renaud 2018-01-23
```
/home/mrenaud/projects/unordered-containers/tests/HashMapProperties.hs:321:9: error:
    • No instance for (Test.QuickCheck.Function.Function Key)
        arising from a use of ‘testProperty’
    • In the expression: testProperty "insertWithKey" pInsertWithKey
      In the second argument of ‘testGroup’, namely
        ‘[testProperty "size" pSize, testProperty "member" pMember,
          testProperty "lookup" pLookup, testProperty "insert" pInsert,
          ....]’
      In the expression:
        testGroup
          "basic interface"
          [testProperty "size" pSize, testProperty "member" pMember,
           testProperty "lookup" pLookup, testProperty "insert" pInsert, ....]
```

I may be using `Fun` wrong, but that's the error when changing the first tuple element from `Int` to `Key`. I'll keep playing around with it.