@treeowl 2018-01-23
Oh, yeah, add the instance for sure. You may need a CoArbitrary one too.

On Jan 22, 2018 11:42 PM, "Matt Renaud" <notifications@github.com> wrote:

/home/mrenaud/projects/unordered-containers/tests/HashMapProperties.hs:321:9:
error:
    • No instance for (Test.QuickCheck.Function.Function Key)
        arising from a use of ‘testProperty’
    • In the expression: testProperty "insertWithKey" pInsertWithKey
      In the second argument of ‘testGroup’, namely
        ‘[testProperty "size" pSize, testProperty "member" pMember,
          testProperty "lookup" pLookup, testProperty "insert" pInsert,
          ....]’
      In the expression:
        testGroup
          "basic interface"
          [testProperty "size" pSize, testProperty "member" pMember,
           testProperty "lookup" pLookup, testProperty "insert" pInsert, ....]

I may be using Fun wrong, but that's the error when changing the first
tuple element from Int to Key. I'll keep playing around with it.

—
You are receiving this because you commented.

Reply to this email directly, view it on GitHub
<https://github.com/tibbe/unordered-containers/pull/177#issuecomment-359674838>,
or mute the thread
<https://github.com/notifications/unsubscribe-auth/ABzi_S4brMtt-9DJTx3zidHvX_00SMK0ks5tNWMogaJpZM4RpAxf>
.