@m-renaud 2018-01-23
Instance issue fixed, the insertWithKey value strictness check is failing though, I'll have to look more closely.