@treeowl 2018-01-23
Recent QuickCheck has apply2 and apply3 or some such; you can upgrade as
required.

On Jan 23, 2018 12:10 AM, "David Feuer" <david.feuer@gmail.com> wrote:

> You can't double curry like that--this is Haskell, not Idris!
>
> On Jan 23, 2018 12:09 AM, "Matt Renaud" <notifications@github.com> wrote:
>
>> Instance issue fixed, the insertWithKey value strictness check is failing
>> though, I'll have to look more closely.
>>
>> —
>> You are receiving this because you commented.
>> Reply to this email directly, view it on GitHub
>> <https://github.com/tibbe/unordered-containers/pull/177#issuecomment-359678187>,
>> or mute the thread
>> <https://github.com/notifications/unsubscribe-auth/ABzi_SJndnW9Z5J3WmN2hbzMuu0B9g4Rks5tNWlsgaJpZM4RpAxf>
>> .
>>
>