@m-renaud 2018-01-23
> You can't double curry like that--this is Haskell, not Idris!

:(

I'm having issues with dependency solving when changing the lower bound. I'll try again tomorrow.