@treeowl 2018-01-23
You can also just write the functions yourself. They're trivial. You'll
find them somewhere in the containers test suite. (I believe QuickCheck
snagged them from me!)

On Jan 23, 2018 12:24 AM, "Matt Renaud" <notifications@github.com> wrote:

> You can't double curry like that--this is Haskell, not Idris!
>
> :(
>
> I'm having issues with dependency solving when changing the lower bound.
> I'll try again tomorrow.
>
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub
> <https://github.com/tibbe/unordered-containers/pull/177#issuecomment-359680064>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/ABzi_ZugsHrieXpcMOv8jv5W2YYJ7JxFks5tNWz9gaJpZM4RpAxf>
> .
>