@m-renaud 2018-01-23
Still getting this strictness property failure:
```
  insertWithKey is value-strict: [Failed]
*** Failed! Falsifiable (after 1 test and 3 shrinks): 
{_->0}
K {unK = 0}
0
fromList []
```

It's weird because `insertWith` uses the same implementation but those tests pass :/