@m-renaud 2018-01-23
Oh duh, the Strict and Lazy modules have their own implementations of insertWithKey, they don't re-export the Base implementation.