@treeowl 2018-01-23
You still need to benchmark old vs. new implementations of `insertWith` (particularly the strict versions) if you want to change that. Or you can just tolerate code duplication.