@treeowl 2018-01-23
But benchmarking is definitely preferred.