@treeowl 2018-01-23
They do indeed. Can you try applying less noise?

On Jan 23, 2018 2:01 PM, "Matt Renaud" <notifications@github.com> wrote:

Old implementation (code duplication):

    benchmarking HashMap/insertWith/String
    time                 2.124 ms   (2.065 ms .. 2.187 ms)
                         0.990 R²   (0.984 R² .. 0.994 R²)
    mean                 1.852 ms   (1.778 ms .. 1.924 ms)
    std dev              222.5 μs   (191.1 μs .. 266.3 μs)
    variance introduced by outliers: 75% (severely inflated)

    benchmarking HashMap/insertWith/ByteString
    time                 2.396 ms   (2.237 ms .. 2.554 ms)
                         0.975 R²   (0.966 R² .. 0.985 R²)
    mean                 2.031 ms   (1.922 ms .. 2.140 ms)
    std dev              331.9 μs   (264.6 μs .. 469.7 μs)
    variance introduced by outliers: 84% (severely inflated)

    benchmarking HashMap/insertWith/Int
    time                 1.252 ms   (1.162 ms .. 1.347 ms)
                         0.976 R²   (0.963 R² .. 0.993 R²)
    mean                 1.137 ms   (1.108 ms .. 1.178 ms)
    std dev              103.3 μs   (73.73 μs .. 165.2 μs)
    variance introduced by outliers: 67% (severely inflated)

New implementation (in terms of insertWithKey):

    benchmarking HashMap/insertWith/String
    time                 1.932 ms   (1.771 ms .. 2.124 ms)
                         0.954 R²   (0.923 R² .. 0.983 R²)
    mean                 1.750 ms   (1.664 ms .. 1.872 ms)
    std dev              301.9 μs   (227.4 μs .. 410.5 μs)
    variance introduced by outliers: 87% (severely inflated)

    benchmarking HashMap/insertWith/ByteString
    time                 2.272 ms   (2.194 ms .. 2.347 ms)
                         0.978 R²   (0.951 R² .. 0.994 R²)
    mean                 1.840 ms   (1.744 ms .. 1.962 ms)
    std dev              331.1 μs   (266.5 μs .. 486.0 μs)
    variance introduced by outliers: 88% (severely inflated)

    benchmarking HashMap/insertWith/Int
    time                 1.321 ms   (1.241 ms .. 1.454 ms)
                         0.955 R²   (0.932 R² .. 0.976 R²)
    mean                 1.365 ms   (1.316 ms .. 1.420 ms)
    std dev              178.3 μs   (150.3 μs .. 230.6 μs)
    variance introduced by outliers: 81% (severely inflated)

The differences appear to be due to noise.

—
You are receiving this because you commented.

Reply to this email directly, view it on GitHub
<https://github.com/tibbe/unordered-containers/pull/177#issuecomment-359894639>,
or mute the thread
<https://github.com/notifications/unsubscribe-auth/ABzi_RuvFp8YfjjVD4JGGcvDflukh10nks5tNix-gaJpZM4RpAxf>
.