@m-renaud 2018-01-23
Updated benchmark results:

**Old (code duplication):**
```
benchmarking HashMap/insertWith/String
time                 1.635 ms   (1.604 ms .. 1.665 ms)
                     0.996 R²   (0.994 R² .. 0.998 R²)
mean                 1.663 ms   (1.636 ms .. 1.695 ms)
std dev              102.2 μs   (81.08 μs .. 130.7 μs)
variance introduced by outliers: 46% (moderately inflated)
             
benchmarking HashMap/insertWith/ByteString
time                 1.634 ms   (1.610 ms .. 1.662 ms)
                     0.997 R²   (0.995 R² .. 0.998 R²)
mean                 1.643 ms   (1.614 ms .. 1.666 ms)
std dev              90.85 μs   (73.72 μs .. 110.7 μs)
variance introduced by outliers: 41% (moderately inflated)
             
benchmarking HashMap/insertWith/Int
time                 1.060 ms   (1.052 ms .. 1.069 ms)
                     0.999 R²   (0.998 R² .. 0.999 R²)
mean                 1.063 ms   (1.054 ms .. 1.076 ms)
std dev              38.48 μs   (29.12 μs .. 54.26 μs)
variance introduced by outliers: 25% (moderately inflated)
```

**New (implemented via insertWithKey):**
```
benchmarking HashMap/insertWith/String
time                 1.687 ms   (1.641 ms .. 1.729 ms)
                     0.994 R²   (0.990 R² .. 0.997 R²)
mean                 1.716 ms   (1.681 ms .. 1.748 ms)
std dev              112.6 μs   (91.79 μs .. 141.6 μs)
variance introduced by outliers: 49% (moderately inflated)
             
benchmarking HashMap/insertWith/ByteString
time                 1.649 ms   (1.621 ms .. 1.679 ms)
                     0.997 R²   (0.996 R² .. 0.998 R²)
mean                 1.650 ms   (1.628 ms .. 1.672 ms)
std dev              69.74 μs   (57.69 μs .. 87.52 μs)
variance introduced by outliers: 28% (moderately inflated)
             
benchmarking HashMap/insertWith/Int
time                 1.081 ms   (1.072 ms .. 1.090 ms)
                     0.999 R²   (0.998 R² .. 0.999 R²)
mean                 1.084 ms   (1.073 ms .. 1.095 ms)
std dev              37.65 μs   (30.82 μs .. 47.29 μs)
variance introduced by outliers: 23% (moderately inflated)
```